+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog of dpnetcdf
 
 - Fix bug when absolute value for year/scenario is None. 
 
+- Enforce lower case map layer names (parameter field) to avoid problems with 
+  GeoServer.
+
 
 0.1 (2013-04-18)
 ----------------